hackerrank7.cpp

+++ hackerrank7.cpp
@@ -7,16 +7,18 @@ int main() {
     int n;
     cin >> n; 
 
-    vector<int> a(n);   
-    vector<int> prefix(n); 
+    vector<long long int> a(n);  
+    vector<long long int> prefix(n); 
 
     for (int i = 0; i < n; i++) {
-        cin >> a[i]; 
+        cin >> a[i];
     }
+
     prefix[0] = a[0];
     for (int i = 1; i < n; i++) {
         prefix[i] = prefix[i - 1] + a[i];
     }
+
     reverse(prefix.begin(), prefix.end());
 
     for (int i = 0; i < n; i++) {
